# Faster `closest()` + benchmark