@tristanls 2018-05-12
On my machine (node v10.1.0)...

Before:
```
:~/.../k-bucket/benchmarks (master %=) 2018-05-12T14:39:09-0500 
$ node ./closest.js 
Seed: ff8a30a71254d3c17d7e4d2c5a63d1025d093935ac7819e9497358925d57e5b5
KBucket.closest: 157.562ms
Memory:  { rss: 31444992,
  heapTotal: 11853824,
  heapUsed: 7236104,
  external: 15028 }
:~/.../k-bucket/benchmarks (master %=) 2018-05-12T14:39:10-0500 
$ node ./closest.js 
Seed: 6d88d89bbd84ddc6009887ad518135f77135236f755121af0632e15fe995e4ea
KBucket.closest: 139.184ms
Memory:  { rss: 30683136,
  heapTotal: 11853824,
  heapUsed: 7354352,
  external: 15156 }
:~/.../k-bucket/benchmarks (master %=) 2018-05-12T14:39:11-0500 
$ node ./closest.js 
Seed: fe0f637be0d7e05d4c69399b5f62419d92e186eabe22e6b61fecc5f16ed39c75
KBucket.closest: 177.060ms
Memory:  { rss: 30576640,
  heapTotal: 11853824,
  heapUsed: 5984104,
  external: 14996 }
:~/.../k-bucket/benchmarks (master %=) 2018-05-12T14:39:12-0500 
$ node ./closest.js 
Seed: 3c9b14ac2ecd12d5f948e677f36f070a96826ca0ee28a870a0319d9712dde2e2
KBucket.closest: 174.138ms
Memory:  { rss: 31293440,
  heapTotal: 11853824,
  heapUsed: 6275944,
  external: 14996 }
```
After
```
:~/.../k-bucket/benchmarks (master *%=) 2018-05-12T14:40:36-0500 
$ node ./closest.js 
Seed: 29c8e1691ffbbaa61a4d33474a08a2aa0dcc3ec672ddf1326fe1ce003ebaf45a
KBucket.closest: 79.442ms
Memory:  { rss: 31031296,
  heapTotal: 11853824,
  heapUsed: 5783904,
  external: 14772 }
:~/.../k-bucket/benchmarks (master *%=) 2018-05-12T14:40:37-0500
$ node ./closest.js 
Seed: fdbd03036cdf88a76209da5c6112775de9256361d85637aa001c4cd877fe6ffd
KBucket.closest: 80.247ms
Memory:  { rss: 31653888,
  heapTotal: 11853824,
  heapUsed: 5412160,
  external: 15028 }
:~/.../k-bucket/benchmarks (master *%=) 2018-05-12T14:40:39-0500
$ node ./closest.js 
Seed: 7db4908b217954a82fecd8b20cb96332a2bdf989975454b350834b3d032e4516
KBucket.closest: 82.668ms
Memory:  { rss: 30846976,
  heapTotal: 11853824,
  heapUsed: 5905384,
  external: 15156 }
:~/.../k-bucket/benchmarks (master *%=) 2018-05-12T14:40:40-0500 
$ node ./closest.js 
Seed: c6255407f0b4bd364c060e33280805d26f7f271c5e5ac0fa13d3aa59233d2382
KBucket.closest: 82.130ms
Memory:  { rss: 31760384,
  heapTotal: 11853824,
  heapUsed: 6077960,
  external: 14548 }
```